Move evaluators to hub (#8)
* Move evaluation.py into parent folder

* Move evaluators into hub

* Change docs for evaluator function import

* Remove evaluation folder
