+++ cognify/__init__.py
@@ -12,8 +12,6 @@ from .frontends.dspy.connector import PredictModel, as_predict
 from .frontends.langchain.connector import RunnableModel, as_runnable
 
 from cognify import llm, optimizer
-from cognify.optimizer.evaluation import metric
-from cognify.optimizer.evaluation.metric import f1_score_set
 
 from cognify.run.evaluate import evaluate, load_workflow
 from cognify.run.optimize import optimize
@@ -25,7 +23,7 @@ from cognify.optimizer import (
     clear_registry,
 )
 
-from cognify.optimizer.evaluation.evaluator import EvaluationResult
+from cognify.optimizer.evaluator import EvaluationResult
 from cognify import _logging
 
 from cognify.hub.search.default import create_search as create_default_search
@@ -35,7 +33,6 @@ __all__ = [
     "llm",
     "optimizer",
     "cogs",
-    "metric",
     "evaluate",
     "load_workflow",
     "optimize",
