+++ cognify/hub/evaluators/match.py
@@ -1,0 +1,2 @@
+def exact_match(pred, label):
+    return 1.0 if pred == label else 0.0
