--- cognify/optimizer/evaluation/metric.py
@@ -1,8 +1,0 @@
-from .f1 import f1_score_set, f1_score_ordered, f1_score_str
-from .llm_judge import llm_judge_generic
-from .code_humaneval import humaneval_evaluator
-
-# Other Metrics
-
-def exact_match(pred, label):
-    return 1.0 if pred == label else 0.0
